requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
